Metrics logging -> warn, limited.
Jira: OTPL-1854

So it turns out that the problem was actually not with the offset
monitor implementation.  What was actually happening in the chat
monitor's case was that the usage was so minimal in QA that only one
number was getting regularly sent to.  So we had data on only one of the
four partitions.  This meant that the other three partitions eventually
lost their offset record (this is expected).  So when the monitor went
to probe Kafka to see how far behind the consumer group was, it wasn't
getting any data for the other three partitions and consequently
declared an error.

This commit updates that log call to be a warning instead of an error.
It also updates the logging to be rate limited at a rate of one per ten
minutes, with a burst capacity of six.