Nulls metrics for "unknown" partitions.
Jira: OTPL-2035

When the offsets is _empty_, we set the lag and offset metrics for each
partition (as reported by the topic size info) to null.

When offsets are known, we ensure we set the individual metrics (e.g.,
partition 0 lag) for the partitions we're _unaware of_ to null.

Note that this implementation makes this determiniation at each poll
time.  This means that this metrics implementation supports reporting
accurate metrics even during *rebalancing*.