Be able to Change the sampling bucket
We started withh the assumption of two models

1. Flip a coin and match the %. This works, but on high volume can over log
2. Token Bucket. We assumed 1 per 10 seconds was as low as folks wanted.

Emanuel has come up with arguments that he wants to log but less commonly.
He can either use 1) with a low % and hope his volume doesn't log too often ;)
or 2) we can expose the ability to define the bucket rate.

I've done the latter